feat: Implement test code of B tree
