+++ src/core/index.ts
@@ -1,1 +1,2 @@
 export * from './BTree';
+export * from './utils';
